requirements.txt
# Full Harbor — consolidated dependencies
# Run: pip install -r requirements.txt

# Core AI / RAG
any-llm-sdk[openai]>=1.10.0
numpy>=1.26.0

# Web scraping & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0

# API layer
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0

# Config
python-dotenv>=1.0.0

# Supabase (for future Supabase integration; optional in CI)
supabase>=2.3.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0

# Linting / type checking (CI)
ruff>=0.3.0
pyright>=1.1.350
